+++ hutool-core/src/main/java/cn/hutool/core/map/CamelCaseLinkedMap.java
@@ -1,8 +1,7 @@
 package cn.hutool.core.map;
 
+import java.util.HashMap;
 import java.util.Map;
-
-import cn.hutool.core.util.StrUtil;
 
 /**
  * 驼峰Key风格的LinkedHashMap<br>
@@ -14,14 +13,44 @@ import cn.hutool.core.util.StrUtil;
  * @param <V> 值类型
  * @since 4.0.7
  */
-public class CamelCaseLinkedMap<K, V> extends CustomKeyLinkedMap<K, V> {
+public class CamelCaseLinkedMap<K, V> extends CamelCaseMap<K, V> {
 	private static final long serialVersionUID = 4043263744224569870L;
 
+	// ------------------------------------------------------------------------- Constructor start
 	/**
 	 * 构造
 	 */
 	public CamelCaseLinkedMap() {
-		super();
+		this(DEFAULT_INITIAL_CAPACITY);
+	}
+
+	/**
+	 * 构造
+	 * 
+	 * @param initialCapacity 初始大小
+	 */
+	public CamelCaseLinkedMap(int initialCapacity) {
+		this(initialCapacity, DEFAULT_LOAD_FACTOR);
+	}
+
+	/**
+	 * 构造
+	 * 
+	 * @param m Map
+	 */
+	public CamelCaseLinkedMap(Map<? extends K, ? extends V> m) {
+		this(DEFAULT_LOAD_FACTOR, m);
+	}
+
+	/**
+	 * 构造
+	 * 
+	 * @param loadFactor 加载因子
+	 * @param m Map
+	 */
+	public CamelCaseLinkedMap(float loadFactor, Map<? extends K, ? extends V> m) {
+		this(m.size(), loadFactor);
+		this.putAll(m);
 	}
 
 	/**
@@ -31,38 +60,7 @@ public class CamelCaseLinkedMap<K, V> extends CustomKeyLinkedMap<K, V> {
 	 * @param loadFactor 加载因子
 	 */
 	public CamelCaseLinkedMap(int initialCapacity, float loadFactor) {
-		super(initialCapacity, loadFactor);
+		super(new HashMap<K, V>(initialCapacity, loadFactor));
 	}
-
-	/**
-	 * 构造
-	 * 
-	 * @param initialCapacity 初始大小
-	 */
-	public CamelCaseLinkedMap(int initialCapacity) {
-		super(initialCapacity);
-	}
-
-	/**
-	 * 构造
-	 * 
-	 * @param m Map
-	 */
-	public CamelCaseLinkedMap(Map<? extends K, ? extends V> m) {
-		super(m);
-	}
-	
-	/**
-	 * 将Key转为驼峰风格，如果key为字符串的话
-	 * 
-	 * @param key KEY
-	 * @return 驼峰Key
-	 */
-	@Override
-	protected Object customKey(Object key) {
-		if (null != key && key instanceof CharSequence) {
-			key = StrUtil.toCamelCase(key.toString());
-		}
-		return key;
-	}
+	// ------------------------------------------------------------------------- Constructor end
 }