+++ hutool-core/src/main/java/cn/hutool/core/map/multi/package-info.java
@@ -1,0 +1,7 @@
+/**
+ * 列表类型值的Map实现
+ * 
+ * @author looly
+ *
+ */
+package cn.hutool.core.map.multi;