+++ hutool-core/src/test/java/cn/hutool/core/util/StrUtilTest.java
@@ -106,8 +106,11 @@ public class StrUtilTest {
 		str = "abcd123";
 		strip = StrUtil.strip(str, null, "567");
 		Assert.assertEquals("abcd123", strip);
+		
+		Assert.assertEquals("", StrUtil.strip("a","a"));
+		Assert.assertEquals("", StrUtil.strip("a","a", "b"));
 	}
-
+	
 	@Test
 	public void stripIgnoreCaseTest() {
 		String str = "abcd123";
@@ -379,6 +382,10 @@ public class StrUtilTest {
 		Assert.assertNull(StrUtil.padPre(null, 10, ' '));
 		Assert.assertEquals("001", StrUtil.padPre("1", 3, '0'));
 		Assert.assertEquals("12", StrUtil.padPre("123", 2, '0'));
+		
+		Assert.assertNull(StrUtil.padPre(null, 10, "AA"));
+		Assert.assertEquals("AB1", StrUtil.padPre("1", 3, "ABC"));
+		Assert.assertEquals("12", StrUtil.padPre("123", 2, "ABC"));
 	}
 	
 	@Test
@@ -386,5 +393,10 @@ public class StrUtilTest {
 		Assert.assertNull(StrUtil.padAfter(null, 10, ' '));
 		Assert.assertEquals("100", StrUtil.padAfter("1", 3, '0'));
 		Assert.assertEquals("23", StrUtil.padAfter("123", 2, '0'));
+		
+		Assert.assertNull(StrUtil.padAfter(null, 10, "ABC"));
+		Assert.assertEquals("1AB", StrUtil.padAfter("1", 3, "ABC"));
+		Assert.assertEquals("23", StrUtil.padAfter("123", 2, "ABC"));
 	}
+	
 }