+++ hutool-crypto/src/main/java/cn/hutool/crypto/SmUtil.java
@@ -34,12 +34,12 @@ public class SmUtil {
 	 * 私钥和公钥同时为空时生成一对新的私钥和公钥<br>
 	 * 私钥和公钥可以单独传入一个，如此则只能使用此钥匙来做加密或者解密
 	 * 
-	 * @param privateKeyBase64 私钥Base64
-	 * @param publicKeyBase64 公钥Base64
+	 * @param privateKeyStr 私钥Hex或Base64表示
+	 * @param publicKeyStr 公钥Hex或Base64表示
 	 * @return {@link SM2}
 	 */
-	public static SM2 sm2(String privateKeyBase64, String publicKeyBase64) {
-		return new SM2(privateKeyBase64, publicKeyBase64);
+	public static SM2 sm2(String privateKeyStr, String publicKeyStr) {
+		return new SM2(privateKeyStr, publicKeyStr);
 	}
 
 	/**