+++ hutool-crypto/src/main/java/cn/hutool/crypto/asymmetric/RSA.java
@@ -69,11 +69,11 @@ public class RSA extends AsymmetricCrypto {
 	 * 私钥和公钥同时为空时生成一对新的私钥和公钥<br>
 	 * 私钥和公钥可以单独传入一个，如此则只能使用此钥匙来做加密或者解密
 	 * 
-	 * @param privateKeyBase64 私钥Base64
-	 * @param publicKeyBase64 公钥Base64
+	 * @param privateKeyStr 私钥Hex或Base64表示
+	 * @param publicKeyStr 公钥Hex或Base64表示
 	 */
-	public RSA(String privateKeyBase64, String publicKeyBase64) {
-		super(ALGORITHM_RSA, privateKeyBase64, publicKeyBase64);
+	public RSA(String privateKeyStr, String publicKeyStr) {
+		super(ALGORITHM_RSA, privateKeyStr, publicKeyStr);
 	}
 
 	/**