+++ hutool-extra/src/main/java/cn/hutool/extra/template/TemplateEngine.java
@@ -5,7 +5,7 @@ package cn.hutool.extra.template;
  * 
  * @author looly
  */
-public interface Engine {
+public interface TemplateEngine {
 	
 	/**
 	 * 获取模板