+++ hutool-extra/src/main/java/cn/hutool/extra/template/TemplateUtil.java
@@ -1,6 +1,6 @@
 package cn.hutool.extra.template;
 
-import cn.hutool.extra.template.engine.EngineFactory;
+import cn.hutool.extra.template.engine.TemplateFactory;
 
 /**
  * 模板工具类
@@ -13,10 +13,10 @@ public class TemplateUtil {
 	/**
 	 * 根据用户引入的模板引擎jar，自动创建对应的模板引擎对象，使用默认配置
 	 * 
-	 * @return {@link Engine}
+	 * @return {@link TemplateEngine}
 	 * @since 4.1.11
 	 */
-	public static Engine createEngine() {
+	public static TemplateEngine createEngine() {
 		return createEngine(new TemplateConfig());
 	}
 
@@ -24,9 +24,9 @@ public class TemplateUtil {
 	 * 根据用户引入的模板引擎jar，自动创建对应的模板引擎对象
 	 * 
 	 * @param config 模板配置，包括编码、模板文件path等信息
-	 * @return {@link Engine}
+	 * @return {@link TemplateEngine}
 	 */
-	public static Engine createEngine(TemplateConfig config) {
-		return EngineFactory.create(config);
+	public static TemplateEngine createEngine(TemplateConfig config) {
+		return TemplateFactory.create(config);
 	}
 }