+++ hutool-extra/src/main/java/cn/hutool/extra/template/engine/TemplateFactory.java
@@ -1,8 +1,10 @@
 package cn.hutool.extra.template.engine;
 
+import com.jfinal.template.Engine;
+
 import cn.hutool.core.util.StrUtil;
-import cn.hutool.extra.template.Engine;
 import cn.hutool.extra.template.TemplateConfig;
+import cn.hutool.extra.template.TemplateEngine;
 import cn.hutool.extra.template.TemplateException;
 import cn.hutool.extra.template.engine.beetl.BeetlEngine;
 import cn.hutool.extra.template.engine.enjoy.EnjoyEngine;
@@ -18,15 +20,15 @@ import cn.hutool.log.StaticLog;
  * @author looly
  *
  */
-public class EngineFactory {
+public class TemplateFactory {
 	/**
 	 * 根据用户引入的模板引擎jar，自动创建对应的模板引擎对象
 	 * 
 	 * @param config 模板配置，包括编码、模板文件path等信息
 	 * @return {@link Engine}
 	 */
-	public static Engine create(TemplateConfig config) {
-		final Engine engine = doCreate(config);
+	public static TemplateEngine create(TemplateConfig config) {
+		final TemplateEngine engine = doCreate(config);
 		StaticLog.debug("Use [{}] Engine As Default.", StrUtil.removeSuffix(engine.getClass().getSimpleName(), "Engine"));
 		return engine;
 	}
@@ -37,7 +39,7 @@ public class EngineFactory {
 	 * @param config 模板配置，包括编码、模板文件path等信息
 	 * @return {@link Engine}
 	 */
-	private static Engine doCreate(TemplateConfig config) {
+	private static TemplateEngine doCreate(TemplateConfig config) {
 		try {
 			return new BeetlEngine(config);
 		} catch (NoClassDefFoundError e) {
@@ -68,6 +70,6 @@ public class EngineFactory {
 		} catch (NoClassDefFoundError e) {
 			// ignore
 		}
-		throw new TemplateException("No template found ! Please add one of [Beetl,Freemarker,Velocity,Rythm] jar to your project !");
+		throw new TemplateException("No template found ! Please add one of template jar to your project !");
 	}
 }