+++ hutool-extra/src/main/java/cn/hutool/extra/tokenizer/engine/ikanalyzer/package-info.java
@@ -1,0 +1,8 @@
+/**
+ * IKAnalyzer分词引擎实现<br>
+ * 项目地址：https://github.com/yozhao/IKAnalyzer
+ * 
+ * @author looly
+ *
+ */
+package cn.hutool.extra.tokenizer.engine.ikanalyzer;