+++ hutool-extra/src/main/java/cn/hutool/extra/tokenizer/package-info.java
@@ -1,0 +1,8 @@
+/**
+ * 中文分词封装<br>
+ * 通过定义统一接口，适配第三方分词引擎
+ * 
+ * @author looly
+ *
+ */
+package cn.hutool.extra.tokenizer;